% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_calculate_delta_r2_for_continuous.R
\name{calculate_delta_r2_for_continuous}
\alias{calculate_delta_r2_for_continuous}
\title{Calculate Delta R2 for Continuous Variables}
\usage{
calculate_delta_r2_for_continuous(
  data = data,
  study = your_study_name,
  continuous_variables = your_continuous_variables,
  control_covariates = 1
)
}
\arguments{
\item{data}{A data frame containing the data with cells, clock, control covariates, and continuous variables of interest. Must include a variable \code{Age}, used for residualization of cells and clocks on age before model fitting.}

\item{study}{A character string specifying the study name, used for naming output files. This argument is required.}

\item{continuous_variables}{A character vector of continuous variables to test for associations with cells and clocks. These variables should be present in \code{data}.}

\item{control_covariates}{A character vector of covariates to include in the regression models.}
}
\value{
A table with the basic model R2, the full model R2, and the difference between them (delta-R2) for each of the user defined \code{continuous_variables}. Also includes number of observations.
}
\description{
This function fits a basic model that uses only user defined \code{control_covariates} (e.g. Array, Batch) to predict clock or cell outcomes and takes the R2. It then fits another model that adds one of the user defined \code{continuous_variables} to the basic model, and takes the R2 from this model. It produces a table with the basic model R2, the full model R2, and the difference between them (delta-R2) for each of the user defined \code{continuous_variables}.
}
