% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calculate_r2_w_age.R
\name{calculate_r2_w_age}
\alias{calculate_r2_w_age}
\title{Calculate R2 for Age, Accounting for Control Covariates}
\usage{
calculate_r2_w_age(
  data = data,
  study = your_study_name,
  control_covariates = 1
)
}
\arguments{
\item{data}{A data frame containing the data with cells, clock, control covariates. Must include a variable \code{Age}.
#' @param study A character string specifying the study name, used for naming output files. This argument is required.}

\item{control_covariates}{A character vector of covariates to include in the regression models.}
}
\value{
A table with the basic model R2 (including only Age), the full model R2 (that includes Age and user defined control covariates), and the difference between them (delta-R2). Also includes number of complete observations.
}
\description{
This function fits a basic model correlates \code{Age} with each cell and clock measure. It then fits another model that adds user defined \code{control_covariates} to the basic model, and takes the R2 from this model. It produces a table with the basic \code{Age} model R2, the full model R2, and the difference between them (delta-R2).
}
