% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/09_exposure_clock_association_with_without_cells.R
\name{exposure_clock_association_with_without_cells}
\alias{exposure_clock_association_with_without_cells}
\title{Perform Stratified Regression with and without Cell Variables and Save Results}
\usage{
exposure_clock_association_with_without_cells(
  cell_clock_df,
  control_covariates = 1,
  categorical_variables,
  study
)
}
\arguments{
\item{control_covariates}{A character vector of control covariates.}

\item{categorical_variables}{A character vector of categorical variables for stratification.}

\item{study}{A character string for naming the output file.}

\item{data}{A data frame containing the data.}
}
\value{
None (saves results to a CSV file).
}
\description{
This function performs regression analysis on clock variables with and without including cell variables,
stratified by categorical variables, and saves the results to a CSV file.
}
