% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stratified_boxplots_clocks.R
\name{stratified_boxplots_clocks}
\alias{stratified_boxplots_clocks}
\title{Stratified Boxplots for Clocks}
\usage{
stratified_boxplots_clocks(
  data = your_data,
  id = your_id,
  study = your_study_name,
  categorical_variables = NULL
)
}
\arguments{
\item{data}{A data frame containing the data with clocks and any categorical variables that will be used to create groups for the boxplots.}

\item{id}{A character string specifying the column name in \code{data} that contains participant or subject IDs (e.g. "id", "participant", etc.). This argument is required.}

\item{study}{A character string specifying the study name, used for naming output files (e.g. "framingham", "whi", etc.). This argument is required.}

\item{categorical_variables}{A character vector of categorical variables on which to stratify the data for boxplots. These variables must be present in \code{data}. Default using \code{NULL} produces plots data without groupings.}
}
\value{
A ggplot object containing the generated boxplots. The function also saves a summary data frame as a CSV file in the "Cells_Clocks_Output/Tables/" directory.
}
\description{
This function generates boxplots for clocks and a summary table of boxplot parameters for the whole dataset. If categorical variables/factors are provided under \code{categorical_variables}, output will also be provided stratified for levels in each categorical variable.
}
