% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_calculate_cell_clock_delta_r2.R
\name{calculate_cell_clock_delta_r2}
\alias{calculate_cell_clock_delta_r2}
\title{Calculate Delta R2 for Each Clock with and without Cells}
\usage{
calculate_cell_clock_delta_r2(
  cell_clock_df,
  control_covariates = 1,
  categorical_variables,
  study
)
}
\arguments{
\item{control_covariates}{A character vector of control covariates.}

\item{categorical_variables}{A character vector of categorical variables to stratify by.}

\item{study}{A character string specifying the study name, used for naming output files.}

\item{data}{A data frame containing the data.}
}
\value{
A data frame with Delta R2 results and saves the results to a CSV file.
}
\description{
This function calculates the Delta R2 for each clock variable, comparing models with and without cell variables.
It also handles categorical variables by stratifying the analysis and calculating Delta R2 for each level of the categorical variable.
}
