% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_delta_r2_for_continuous.R
\name{calculate_delta_r2_for_continuous}
\alias{calculate_delta_r2_for_continuous}
\title{Calculate Delta R-squared for Continuous Variables}
\usage{
calculate_delta_r2_for_continuous(
  data,
  study,
  cell_types,
  all_clocks,
  continuous_variables,
  control_covariates,
  age_column,
  output_dir = NULL,
  save_results = FALSE
)
}
\arguments{
\item{data}{A data frame containing cell types, clock measures, age, continuous variables,
and control covariates.}

\item{study}{A character string specifying the study name, used for naming output files.}

\item{cell_types}{A character vector of column names in \code{data} that contain cell type measurements.}

\item{all_clocks}{A character vector of column names in \code{data} that contain clock measurements.}

\item{continuous_variables}{A character vector of column names in \code{data} representing
the continuous variables to be analyzed.}

\item{control_covariates}{A character vector of column names in \code{data} to use as
additional control covariates (Batch, Array, Plate, Row, Column etc.) in the regression models.}

\item{age_column}{A character string specifying the name of the age column in \code{data}.}

\item{output_dir}{A character string specifying the output directory. If \code{NULL}, the current
working directory is used. Default is \code{NULL.}}

\item{save_results}{Logical; if TRUE, results are saved as a CSV file. Default is FALSE.}
}
\value{
A data frame with columns:
\item{Cell_or_Clock_Outcome}{Name of the age-adjusted cell type or clock measure}
\item{Continuous_Variable}{Name of the continuous variable added to the model}
\item{R2_no_continuous_variable}{R-squared value for the model without the continuous variable}
\item{R2_w_continuous_variable}{R-squared value for the model with the continuous variable}
\item{Delta_R2}{Difference between the two R-squared values}
\item{n_obs}{Number of complete observations used in the model}
}
\description{
This function calculates R-squared values for the relationship between
user-defined \code{control_covariates} and age-adjusted cell type and clock measures,
both with and without the additional of user-defined \code{continuous_variables}.
It produces a table with the R2 values for a basic model (\code{control_covariates} only), a full model
(\code{control_covariates} plus \code{continuous_variables}), and the difference between them (delta-R2).
}
\examples{
\dontrun{
result <- calculate_delta_r2_for_continuous(
  data = data,
  study = "my_study",
  cell_types = c("Neutrophils", "Monocytes"),
  all_clocks = c("PCGrimAge", "PCPhenoAge"),
  continuous_variables = c("BMI", "Cholesterol"),
  control_covariates = c("Sex", "Smoking"),
  age_column = "Age",
  save_results = TRUE
)
}

}
