% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_exposure_cell_clock_associations.R
\name{exposure_cell_clock_associations}
\alias{exposure_cell_clock_associations}
\title{Perform Regression, Standardization, and Delta R2 Calculations}
\usage{
exposure_cell_clock_associations(
  cell_clock_df,
  control_covariates = 1,
  categorical_variables,
  study
)
}
\arguments{
\item{control_covariates}{A character vector of control covariates.}

\item{categorical_variables}{A character vector of categorical variables for regression.}

\item{study}{A character string for naming the output file.}

\item{data}{A data frame containing the data.}
}
\value{
None (saves results to a CSV file).
}
\description{
This function performs the following steps:
\enumerate{
\item Creates standardized residuals for each specified column using an existing function.
\item Runs linear regressions for each combination of clock and cell residuals, with specified covariates.
\item Saves the regression results to a CSV file.
}
}
