% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_boxplots_cells.R
\name{stratified_boxplots_cells}
\alias{stratified_boxplots_cells}
\title{Generate Stratified Boxplots for Cell Types}
\usage{
stratified_boxplots_cells(
  data,
  id,
  study,
  cell_types,
  highlighted_cell_types = NULL,
  categorical_variables = NULL,
  colors = NULL,
  output_dir = NULL,
  save_plots = FALSE,
  save_summaries = FALSE
)
}
\arguments{
\item{data}{A data frame containing cell type data and optional categorical variables.}

\item{id}{A string specifying the column name for participant/subject IDs.}

\item{study}{A string specifying the study name, used for naming output files.}

\item{cell_types}{A character vector of column names present in \code{data} that contain the cell type data.}

\item{highlighted_cell_types}{A character vector of cell types to be highlighted
with an adjusted y-axis scale. Must be a subset of \code{cell_types}. Default is NULL
where no cell types are highlighted. Neutrophil cell counts are typically highlighted
as they greatly exceed those of other cell types.}

\item{categorical_variables}{A character vector of categorical variables for
stratification. These must be present in \code{data}. Default using \code{NULL} produces
data without groupings.}

\item{colors}{A character vector of colors that can be used to replace
the default color palette. If an insufficient number of colors are provided for the
levels of a categorical variable, default colors or a color ramp will be used.}

\item{output_dir}{A string specifying the output directory. If \code{NULL}, the current
working directory is used.}

\item{save_plots}{Logical; if TRUE, plots are saved in 'cellclockR_output/Plots/'.}

\item{save_summaries}{Logical; if TRUE, summaries are saved in 'cellclockR_output/Summaries/'.}
}
\value{
A list containing the ggplot objects and summary data frames for each categorical variable.
}
\description{
This function creates boxplots for cell types and a summary table of boxplot
parameters for the entire dataset. If categorical variables are provided,
the output will be stratified by levels in each categorical variable.
}
\examples{
\dontrun{
result <- stratified_boxplots_cells(
  data = data,
  id = "participant_id",
  study = "my_study",
  cell_types = c("Neutrophils", "Basophils", "Monocytes"),
  highlighted_cell_types = c("Neutrophils")
  categorical_variables = c("Sex", "AgeGroup"),
  save_plots = TRUE,
  save_summaries = TRUE
)
}

}
