% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_boxplots_clocks.R
\name{stratified_boxplots_clocks}
\alias{stratified_boxplots_clocks}
\title{Generate Stratified Boxplots for Clocks}
\usage{
stratified_boxplots_clocks(
  data,
  id,
  study,
  all_clocks,
  highlighted_clocks = NULL,
  categorical_variables = NULL,
  colors = NULL,
  output_dir = NULL,
  save_plots = FALSE,
  save_summaries = FALSE
)
}
\arguments{
\item{data}{A data frame containing clock data and optional categorical variables.}

\item{id}{A string specifying the column name for participant/subject IDs.}

\item{study}{A string specifying the study name, used for naming output files.}

\item{all_clocks}{A character vector of column names present in \code{data} that contain the clock data.}

\item{categorical_variables}{A character vector of categorical variables for
stratification. These must be present in \code{data}. Default using \code{NULL} produces
data without groupings.}

\item{colors}{A character vector of colors that can be used to replace
the default color palette. If an insufficient number of colors are provided for the
levels of a categorical variable, default colors or a color ramp will be used.}

\item{output_dir}{A string specifying the output directory. If \code{NULL}, the current
working directory is used.}

\item{save_plots}{Logical; if TRUE, plots are saved in 'cellclockR_output/Plots/'.}

\item{save_summaries}{Logical; if TRUE, summaries are saved in 'cellclockR_output/Summaries/'.}

\item{highlighted_clock_types}{A character vector of clocks to be highlighted
with an adjusted y-axis scale. Must be a subset of \code{all_clocks}. Default is NULL
where no clocks are highlighted.}
}
\value{
A list containing the ggplot objects and summary data frames for each categorical variable.
}
\description{
This function creates boxplots for clocks and a summary table of boxplot
parameters for the entire dataset. If categorical variables are provided,
the output will be stratified by levels in each categorical variable.
}
\examples{
\dontrun{
result <- stratified_boxplots_clocks(
  data = data,
  id = "participant_id",
  study = "my_study",
  all_clocks = c("DunedinPACE", "PCPhenoAge", "PCGrimAge"),
  highlighted_cell_types = c("DunedinPACE")
  categorical_variables = c("Sex", "AgeGroup"),
  save_plots = TRUE,
  save_summaries = TRUE
)
}

}
